모델 추가

+++ src/main/java/com/KidLove/atch/vo/AtchVO.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AtchVO {
 
-	private Integer atchNo;
+	private Long atchNo;
 	private String atchOrginlnm;
 	private String atchExtsn;
 	private String atchCode;
